refactor(runtime): split history helpers and limits
