+++ crates/backend/src/agent_runtime.rs
@@ -4,6 +4,7 @@ mod connection;
 mod events;
 mod handoff;
 mod history;
+mod limits;
 mod load;
 mod operations;
 pub(crate) mod plugin_agent;
@@ -33,6 +34,7 @@ use crate::app_event::AppEventPublisher;
 use attach::RebuiltBinding;
 use handoff::HandoffDebt;
 use history::{LocalHistoryClock, RecordOutcome, SessionRecorder};
+use limits::*;
 use load::UnreadableHistory;
 pub use operations::AgentRuntime;
 pub use stream::SessionEventStream;
@@ -76,13 +78,6 @@ use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::{Arc, RwLock};
 use std::time::Duration;
 use tokio::sync::{mpsc, oneshot};
-
-const INITIALIZE_TIMEOUT: Duration = Duration::from_secs(15);
-const SESSION_SETUP_TIMEOUT: Duration = Duration::from_secs(30);
-const CANCELLATION_GRACE: Duration = Duration::from_secs(5);
-const PROMPT_INACTIVITY_TIMEOUT: Duration = Duration::from_secs(5 * 60);
-const CONTRACT_QUEUE_CAPACITY: usize = 256;
-const MAX_PROMPT_BYTES: usize = 16 * 1024 * 1024;
 
 /// Repairs live sessions after the agent process behind them was replaced.
 ///
